Removes onageaent_utils.go
