+++ controllers/oneagent/oneagent_controller_test.go
@@ -11,6 +11,7 @@ import (
 
 	dynatracev1alpha1 "github.com/Dynatrace/dynatrace-operator/api/v1alpha1"
 	"github.com/Dynatrace/dynatrace-operator/controllers"
+	"github.com/Dynatrace/dynatrace-operator/controllers/activegate/reconciler/statefulset"
 	"github.com/Dynatrace/dynatrace-operator/dtclient"
 	"github.com/Dynatrace/dynatrace-operator/logger"
 	"github.com/Dynatrace/dynatrace-operator/scheme"
@@ -571,4 +572,132 @@ func TestServiceAccountName(t *testing.T) {
 		podSpecs = newPodSpecForCR(&instance, &instance.Spec.ClassicFullStack, ClassicFeature, true, log, testClusterID)
 		assert.Equal(t, testName, podSpecs.ServiceAccountName)
 	})
+}
+
+func TestOneAgent_Validate(t *testing.T) {
+	oa := newOneAgent()
+	assert.Error(t, validate(oa))
+	oa.Spec.APIURL = "https://f.q.d.n/api"
+	assert.NoError(t, validate(oa))
+}
+
+func TestMigrationForDaemonSetWithoutAnnotation(t *testing.T) {
+	oaKey := metav1.ObjectMeta{Name: "my-oneagent", Namespace: "my-namespace"}
+
+	ds1 := &appsv1.DaemonSet{ObjectMeta: oaKey}
+
+	ds2, err := newDaemonSetForCR(consoleLogger, &dynatracev1alpha1.DynaKube{ObjectMeta: oaKey}, &dynatracev1alpha1.FullStackSpec{}, "classic", "cluster1")
+	assert.NoError(t, err)
+	assert.NotEmpty(t, ds2.Annotations[statefulset.AnnotationTemplateHash])
+
+	assert.True(t, hasDaemonSetChanged(ds1, ds2))
+}
+
+func TestHasSpecChanged(t *testing.T) {
+	runTest := func(msg string, exp bool, mod func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube)) {
+		t.Run(msg, func(t *testing.T) {
+			key := metav1.ObjectMeta{Name: "my-oneagent", Namespace: "my-namespace"}
+			oldInstance := dynatracev1alpha1.DynaKube{ObjectMeta: key}
+			newInstance := dynatracev1alpha1.DynaKube{ObjectMeta: key}
+
+			mod(&oldInstance, &newInstance)
+
+			ds1, err := newDaemonSetForCR(consoleLogger, &oldInstance, &oldInstance.Spec.ClassicFullStack, "classic", "cluster1")
+			assert.NoError(t, err)
+
+			ds2, err := newDaemonSetForCR(consoleLogger, &newInstance, &newInstance.Spec.ClassicFullStack, "classic", "cluster1")
+			assert.NoError(t, err)
+
+			assert.NotEmpty(t, ds1.Annotations[statefulset.AnnotationTemplateHash])
+			assert.NotEmpty(t, ds2.Annotations[statefulset.AnnotationTemplateHash])
+
+			assert.Equal(t, exp, hasDaemonSetChanged(ds1, ds2))
+		})
+	}
+
+	runTest("no changes", false, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {})
+
+	runTest("image added", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		new.Spec.OneAgent.Image = "docker.io/dynatrace/oneagent"
+	})
+
+	runTest("image set but no change", false, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.OneAgent.Image = "docker.io/dynatrace/oneagent"
+		new.Spec.OneAgent.Image = "docker.io/dynatrace/oneagent"
+	})
+
+	runTest("image removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.OneAgent.Image = "docker.io/dynatrace/oneagent"
+	})
+
+	runTest("image changed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.OneAgent.Image = "registry.access.redhat.com/dynatrace/oneagent"
+		new.Spec.OneAgent.Image = "docker.io/dynatrace/oneagent"
+	})
+
+	runTest("argument removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.Args = []string{"INFRA_ONLY=1", "--set-host-property=OperatorVersion=snapshot"}
+		new.Spec.ClassicFullStack.Args = []string{"INFRA_ONLY=1"}
+	})
+
+	runTest("argument changed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.Args = []string{"INFRA_ONLY=1"}
+		new.Spec.ClassicFullStack.Args = []string{"INFRA_ONLY=0"}
+	})
+
+	runTest("all arguments removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.Args = []string{"INFRA_ONLY=1"}
+	})
+
+	runTest("resources added", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		new.Spec.ClassicFullStack.Resources = newResourceRequirements()
+	})
+
+	runTest("resources removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.Resources = newResourceRequirements()
+	})
+
+	runTest("resources removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.Resources = newResourceRequirements()
+	})
+
+	runTest("priority class added", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		new.Spec.ClassicFullStack.PriorityClassName = "class"
+	})
+
+	runTest("priority class removed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.PriorityClassName = "class"
+	})
+
+	runTest("priority class set but no change", false, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.PriorityClassName = "class"
+		new.Spec.ClassicFullStack.PriorityClassName = "class"
+	})
+
+	runTest("priority class changed", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		old.Spec.ClassicFullStack.PriorityClassName = "some class"
+		new.Spec.ClassicFullStack.PriorityClassName = "other class"
+	})
+
+	runTest("dns policy added", true, func(old *dynatracev1alpha1.DynaKube, new *dynatracev1alpha1.DynaKube) {
+		new.Spec.ClassicFullStack.DNSPolicy = corev1.DNSClusterFirst
+	})
+}
+
+func newResourceRequirements() corev1.ResourceRequirements {
+	return corev1.ResourceRequirements{
+		Limits: corev1.ResourceList{
+			"cpu":    parseQuantity("10m"),
+			"memory": parseQuantity("100Mi"),
+		},
+		Requests: corev1.ResourceList{
+			"cpu":    parseQuantity("20m"),
+			"memory": parseQuantity("200Mi"),
+		},
+	}
+}
+
+func parseQuantity(s string) resource.Quantity {
+	q, _ := resource.ParseQuantity(s)
+	return q
 }
